requirements.txt
langchain
langchain_community
ollama
python-dotenv
fastapi
uvicorn
pydantic
langchain_core